Report correct durations using periods
Change f082711 introduce a "extract_to" parameter, and an additional
check to not report machines that fall outside that interval. However,
this prevents long running VMs to be published. The accounting system
expects absolute durations (for wall and CPU) from the start of the
record to its end or to the reporting date if it has not ended, that is:

    wall = (end or report time) - start time

However, we were doing it simply wrong. Moreover, we were reporting wall
time as the CPU time, and this is not true. From an IaaS perspective,
CPU time (or vCPU time) should be (wall * cpus) as a CPU allocated is a
CPU being used.