Use correct server's start time
When a machine is resized, rescued or evacuated the server's start_time
(and the usages' start_time) is reset and it does not reflect anymore
the time the instance was created. Instead of relying on that time, we
move to use the server.created time.

This solves the issue of negative accounting times, that happen when an
instance has its start_time reset.

Fixes #56