extractor: new per-project base extractor
This change creates a new base extractor that is aware of the project
that we are getting the accounting for. This way the code gets much
simpler, as it is not needed to keep track of projects, clients, etc.
within the extrating methods. As a side effect, this change breaks the
old ceilometer extractor.